pjd-fstest-20090130-RC/tests/xacl/01.t:13
chmod should preserve acls.
Previously chmod just deleted acls, but of course it's not enough.
Now it checks if the file has an acl and if so it opens it and updates the
required fields. The use of __acl_to_xattr was quite heroic, it was clearly not
intended in libacl.