buffers: use a global lock for all operations
As usual I was too optimistic for the usage of locks. Even with fast
operations, there can still be an update on a buffer while another thread is
busy freeing it. It's very rare, and very hard to reproduce, but it's possible.
This global lock used everytime the buffers are used should make this safe.
And it's global and not attached to each file because buffers for a given
physical file are not always accessible through the current opened file
information. It shouldn't make any speed difference most of the time anyway.