re-enable the -O option of zfs mount
this option must be passed to enable a mount even when the dir is not empty on
solaris. In linux it's here for compatibility only, but without it zfs-fuse
tests directly if the directory is empty and does not trust mount about that,
effectively blocking the fuse "nonempty" mount option.
For some reason when you have a few nested filesystems, they are not always
mounted in the right order, and this option is useful in this case.