Revert "resurrect zfs_zget param zget_unlinked"
This reverts commit ea6df50db0b4cfc167925ed452d2840b14ecb103.

Conflicts:

	src/zfs-fuse/zfs_acl.c
	src/zfs-fuse/zfs_dir.c
	src/zfs-fuse/zfs_operations.c
	src/zfs-fuse/zfs_vnops.c

zfs_zget: tune z_unlinked test

it's a wonder that this thing works correctly in solaris.
Maybe getattr is cached there and never asks for this data here.
With fuse when an inode is deleted and is still opened by something
the next getattr call will come here.
So the test is changed to return ENOENT only if there are no opened
file on the inode. So far I couldn't crash this, it's at least much
lighter than the old fix, and should be better.