Much faster unmounting
these sync() calls were here to prevent zpool export to fail when having a tree
of subvolumes inside. Well I just replaced this by an empty loop when needed,
and couldn't make it fail so far. Use with caution, it might fail in some
specific way, but it's very useful when having 1000 clones to unmount !